Bring back files

+++ put.io/File.swift
@@ -11,8 +11,8 @@ import CoreData
 
 @objc(File)
 class File: NSManagedObject {
-    @NSManaged var id: Int
-    @NSManaged var parent_id: Int
+    @NSManaged var id: Double
+    @NSManaged var parent_id: Double
     @NSManaged var name: String
     @NSManaged var content_type: String
     @NSManaged var isFolder: Bool
@@ -23,10 +23,10 @@ class File: NSManagedObject {
         if let _name = json["name"] as? String {
             name = _name
         }
-        if let _id = json["id"] as? Int {
+        if let _id = json["id"] as? Double {
             id = _id
         }
-        if let _parent_id = json["parent_id"] as? Int {
+        if let _parent_id = json["parent_id"] as? Double {
             parent_id = _parent_id
 //            parent = FileStore.getFile(parent_id)
         }
